[Gold II] Title: 피보나치 수 3, Time: 228 ms, Memory: 29520 KB -BaekjoonHub

+++ 백준/Gold/2749. 피보나치 수 3/피보나치 수 3.java
@@ -3,7 +3,7 @@ import java.util.Scanner;
 public class Main {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
-		long n = sc.nextLong() % 1500000;
+		int n = (int) (sc.nextLong() % 1500000);
 		sc.close();
 
 		long[] arr = new long[1500001];
@@ -13,6 +13,6 @@ public class Main {
 		for (int i = 2; i < 1500001; i++)
 			arr[i] = (arr[i - 2] + arr[i - 1]) % 1000000;
 
-		System.out.println(arr[(int) n]);
+		System.out.println(arr[n]);
 	}
 }
